Rename file and fix linter offenses
